move scope util out of katsu utils
