+++ bento_beacon/endpoints/info.py
@@ -1,13 +1,14 @@
 from flask import Blueprint, current_app
 from ..authz.middleware import authz_middleware
 from ..utils.beacon_response import beacon_info_response
-from ..utils.katsu_utils import katsu_get, katsu_datasets, verify_request_project_scope
-from ..utils.scope import scoped_route_decorator_for_blueprint
+from ..utils.katsu_utils import katsu_get, katsu_datasets
+from ..utils.scope import scoped_route_decorator_for_blueprint, verify_request_project_scope
 
 info = Blueprint("info", __name__)
 route_with_optional_project_id = scoped_route_decorator_for_blueprint(info)
 
-
+# this blueprint does not check project / dataset permissions or take requests
+# but needs to verify project ids in the url path 
 @info.before_request
 async def before_info_request():
     await verify_request_project_scope()
