+++ bento_beacon/utils/scope.py
@@ -1,4 +1,10 @@
-# TODO: futureproof: don't ever treat the string "network" as a project id
+from flask import request
+from .katsu_utils import katsu_projects
+from .exceptions import InvalidQuery
+
+MESSAGE_FOR_TOO_MANY_DATASETS = "'datasetIds' field currently cannot take more than one dataset id"
+
+
 def scoped_route_decorator_for_blueprint(blueprint):
     """
     Generates a decorator equivalent to two flask "@route" decorators, one with a project_id prefix, and one without.
@@ -25,4 +31,10 @@ def scoped_route_decorator_for_blueprint(blueprint):
     return scoped_route
 
 
-MESSAGE_FOR_TOO_MANY_DATASETS = "'datasetIds' field currently cannot take more than one dataset id"
+# used by info endpoints that don't check censorship settings
+async def verify_request_project_scope():
+    view_args = request.view_args if request.view_args else {}
+    project_id = view_args.get("project_id")
+    project_ids = [p["identifier"] for p in (await katsu_projects()).get("results", [])]
+    if project_id is not None and project_id not in project_ids:
+        raise InvalidQuery(f"No project found with id {project_id}")
